requirements.txt
asgiref==3.2.10
Django==3.1.1
numpy==1.19.1
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
